Fix issue #001
The error was due to the assignment of the .invisible class to the phone
menu that actually was not appended anywhere.
Instead of add/remove class to the phone menu, it was necessary to
append/remove the div containing the phone menu.

Note: new issue (#002) concerning phone links that don't close the phone menu
once clicked as they're expected to do